draft: split app and models to separate files
